+++ src/Illuminate/Testing/Concerns/TestDatabases.php
@@ -42,7 +42,7 @@ trait TestDatabases
             $databaseTraits = [
                 Testing\DatabaseMigrations::class,
                 Testing\DatabaseTransactions::class,
-                Testing\DatabaseTruncates::class,
+                Testing\DatabaseTruncation::class,
                 Testing\RefreshDatabase::class,
             ];
